Fix MSVS warnings.

+++ Source/PluginProcessor.h
@@ -82,7 +82,7 @@ private:
     std::vector<double> yn_1;
     std::vector<double> yn_2;
 
-    double denormal {1.0e-16}; // use to keep mantissa from dropping below 1.xxx
+    float denormal {1.0e-16}; // use to keep mantissa from dropping below 1.xxx
                                // see http://www.earlevel.com/main/2012/12/03/a-note-about-de-normalization/
                                // also see Falco's DspFilters MathSupplement.h re: 1e-8?
